Finished the bisection code
I am pretty sure that my bisection code works. I have checked it with
another code that determines the number of events kept when both algs
are passed at a certain threshold.
Bisection determines one point of the combined events kept function in
terms of the fraction individually kept by alg A or alg B.
Given a fraction to keep of A, determines the fraction to keep of B such
that 10^(-4) events are kept on combined alg. So this will determine
points on the PPF of threshold of A versus threshold of B.