Finally got the threeEfficiencies.c to work properly
However, there is a lot of copied and pasted code. I would like to be
able to encapsulate different parts of the code in separate macros and
then to call them separately so the code is more readable and compact.
However, I am having issues with compiling into an ACLiC library, and am
receiving errors when I try to use the full C++ compiler.
Also, I want to know how I can return multiple values in a macro, and
what the accepted method of calling other macros from within macros is.
I have emailed Professor Cranmer to see if he has any input.